+++ backend/api/enfiletesbaskets/src/main/java/com/enfiletesbaskets/enfiletesbaskets/models/UserModel.java
@@ -17,7 +17,7 @@ import java.util.List;
 public class UserModel {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false, unique = true)
     private String email;
@@ -56,15 +56,15 @@ public class UserModel {
 
     public UserModel() {}
 
-    public UserModel(Integer id) {
+    public UserModel(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
