+++ backend/api/enfiletesbaskets/src/main/java/com/enfiletesbaskets/enfiletesbaskets/repositories/ClassRepository.java
@@ -12,18 +12,18 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClassRepository extends JpaRepository<ClassModel, Integer> {
+public interface ClassRepository extends JpaRepository<ClassModel, Long> {
     @Query(value = "SELECT c.* " +
                    "FROM Class c " +
                    "JOIN Class_Courses cc ON c.id = cc.class_id " +
                    "JOIN User_Courses uc ON cc.course_id = uc.course_id " +
                    "WHERE uc.user_id = :userId", nativeQuery = true)
-    List<ClassModel> findSubscribedClassesByUserId(@Param("userId") Integer userId);
+    List<ClassModel> findSubscribedClassesByUserId(@Param("userId") Long userId);
 
     @Query("SELECT c FROM ClassModel c WHERE c.password = :password")
     Optional<ClassModel> findByPassword(@Param("password") String password);
 
     @Modifying
     @Query(value = "INSERT INTO Class_Courses (class_id, course_id) VALUES (:classId, :courseId)", nativeQuery = true)
-    void addCourseToClass(@Param("classId") Integer classId, @Param("courseId") Integer courseId);
+    void addCourseToClass(@Param("classId") Long classId, @Param("courseId") Long courseId);
 }
